Here is an updated version with all classes in their own files.
