+++ src/dev/java/game/Main.java
@@ -133,81 +133,4 @@ public class Main {
         }
     }
 
-    public static class Item {
-        private String name;
-        private String description;
-
-        // Detailed description that is given when inspected
-        private String inspection;
-
-        // Keeps track of if item has been observed yet
-        private boolean observed;
-
-        public Item(String name, String description, String inspection) {
-            this.name = name;
-            this.description = description;
-            this.inspection = inspection;
-            this.observed = false;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public String getDescription() {
-            return description;
-        }
-
-        public void setObserved(boolean observed) {
-            this.observed = observed;
-        }
-
-        public boolean isObserved() {
-            return observed;
-        }
-
-        public String getInspection() {
-            return inspection;
-        }
-        @Override
-        public String toString() {
-            return name + ": " + description;
-        }
-    }
-
-    public static class Room {
-        // this is the logical representation of the "walls" of each room
-        private EnumMap<Direction, Item> walls; // map of directions to items. syntax: <key, value>
-
-        // Keeps track of items in separate map not associated with direction
-        // This way player can inspect item without specifying direction
-        // TODO: Are we adding items to this map after player "discovers" them via look?
-        private Map<String, Item> items = new HashMap<>();
-        // this string holds the name of the room, which is set at instantiation
-        // We could also maybe think about pre-baking constructors for our rooms, that way they're built at runtime and just need to be loaded in on the fly
-        private String name = "Undefined Room Name";
-
-        public Room() {
-            walls = new EnumMap<>(Direction.class); // initializes map
-        }
-
-        public void setItem(Direction direction, Item item) {
-            walls.put(direction, item); // adds item to specified direction
-            items.put(item.name, item); // adds item to item list
-        }
-
-        public Item getItemAtDirection(Direction direction) {
-            return walls.get(direction); // gets item at the specified direction
-        }
-        public void setName(String name) {
-            this.name = name;
-        }
-        public String getName() {
-            return name;
-        }
-
-        public Map<String, Item> getItems() {
-            return items;
-        }
-    }
 }
